refactoring into logical units of thought
